Address PR comments

+++ service/src/main/java/bio/terra/profile/service/spendreporting/azure/model/SpendDataItem.java
@@ -1,4 +1,6 @@
 package bio.terra.profile.service.spendreporting.azure.model;
 
+import java.math.BigDecimal;
+
 public record SpendDataItem(
-    String resourceType, Double cost, String currency, SpendCategoryType spendCategoryType) {}
+    String resourceType, BigDecimal cost, String currency, SpendCategoryType spendCategoryType) {}
